fix: pin “Copy” button to bottom of panel
The style of the `ActionBar` that includes the “Copy” button now visually matches the Storybook-provided addons, where it always appears at the very bottom of the panel.

Previously, the button would appear just below the bottom of the content in cases where the HTML snippet did not vertically fill the panel.