fix: ensure code is shown on first render
When using `useAddonState`, the code snippet would not appear on the first load of the Storybook page; it would only appear after changing to a different story. From that point on, it worked as expected.

Since this state is not shared outside of this component, it shouldn’t be necessary to use the `useAddonState` hook; just using the normal React `useState` hook should work fine, and seems to fix the problem.